requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
pydantic>=2.0
pydantic-settings>=2.0
pyyaml>=6.0
python-ulid>=3.0
python-multipart>=0.0.9
jinja2>=3.1
pillow>=10.0
opencv-python-headless>=4.9
numpy>=1.26
websockets>=12.0
httpx>=0.27
openai>=1.30
anthropic>=0.30
langgraph>=0.2
langchain-core>=0.3
scikit-image>=0.22
